Separated numbers into its own component
